changed general data structure and main_photo save

+++ auth/middleware.go
@@ -35,7 +35,7 @@ func AuthMiddleware(publicKey *rsa.PublicKey) func(http.Handler) http.Handler {
 
 			ctx := r.Context()
 			ctx = context.WithValue(ctx, "claims", models.AuthContext{
-				UserID: claims["user_id"].(int64),
+				UserID: int64(claims["user_id"].(float64)),
 				Role:   claims["role"].(string),
 			})
